Support "canvassize" configuration command
This approximates the behavior of the Hall C onlineGUI, which uses the
canvassize to set the size of the embedded canvas and hardcodes the sizes of the
surrounding GUI elements (sidebar etc.). Panguin scales all these elements
depending on the main window size. Here, we scale the canvassize to the main
window size based on the scale factors in OnlineGUI::CreateGUI.

Also, rearrange some fields in OnlineConfig to pack the class more
efficiently.